Use the SECRET_KEY from the environment variable
The /etc/tower/SECRET_KEY is used as the ansible tower encryption key
for sensitive information stored in the database.

This key is also installed when the rpm is installed (which is to
say, into the image). In order for the key not to be the same across
all images we expect it to be specified as an environment variable
at run time.